apps/api/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlmodel==0.0.24
SQLAlchemy==2.0.43
pydantic-settings==2.11.0
httpx==0.28.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
websockets==15.0.1
pytest==8.4.1
